+++ src/context.rs
@@ -175,6 +175,10 @@ pub trait VariableMap {
     fn get_rng(&self) -> &RefCell<Pcg32>;
 }
 
+pub trait ConfigView {
+    fn config(&self) -> &TransformConfig;
+}
+
 pub trait ContextView: ElementMap + VariableMap {}
 
 impl ElementMap for TransformerContext {
@@ -285,6 +289,12 @@ impl VariableMap for TransformerContext {
 
     fn get_rng(&self) -> &RefCell<Pcg32> {
         &self.rng
+    }
+}
+
+impl ConfigView for TransformerContext {
+    fn config(&self) -> &TransformConfig {
+        &self.config
     }
 }